requirement.txt
#python=3.6.13

pillow==8.4.0
pandas==1.1.5
seaborn==0.11.2
scikit-learn==0.24.2
scikit-image==0.17.2
imutils==0.5.4
tensorflow==1.15.0
keras==2.2.4
opencv_contrib_python==3.4.5.20
playsound==1.2.2
pyttsx3==2.98
wordsegment==1.3.1
pyenchant==3.2.2
autocomplete==0.0.104